requirements.txt
django-mptt
django-filer
djangorestframework
django-filter
markdown
